refactor StationView
