+++ LuasWatch Watch App/Views/LuasTextView.swift
@@ -1,0 +1,17 @@
+//
+//  Created by Roland Gropmair on 10/02/2024.
+//  Copyright © 2024 mApps.ie. All rights reserved.
+//
+
+import SwiftUI
+
+struct LuasTextView: View {
+    var text: String
+
+    var body: some View {
+        Text(text)
+            .multilineTextAlignment(.center)
+            .frame(idealHeight: .greatestFiniteMagnitude)
+            .padding(.horizontal, 8)
+    }
+}
